requirements.txt
xhtml2pdf==0.2.5
django-environ==0.4.5
environ==1.0
Django==3.2.4
timeago==1.0.15
django_heroku==0.3.1
hashids==1.3.1
pycryptodome==3.10.1
gunicorn==20.1.0
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
uuid==1.30
Werkzeug==1.0.1
whitenoise==5.1.0